feat: loginRequired 적용

+++ src/controllers/post.controller.ts
@@ -1,21 +1,31 @@
-import { NextFunction, Request, Response } from 'express';
-import postService from '../services/post.service';
-import { PostType, PostParamsType, PostInfoType } from '../types/post';
 import Joi from 'joi';
 
+import { NextFunction, Request, Response } from 'express';
+import postService from '../services/post.service';
+import ApiError from '../utils/ApiError';
+
+//type
+import { PostParamsType, PostInfoType } from '../types/postType';
+import { UserIdReqType } from '../types/loginRequiredType';
+
 const postReqValidator = Joi.object({
-  title: Joi.string().required().trim(),
-  description: Joi.string().required(),
+  title: Joi.string().trim(),
+  description: Joi.string().trim(),
 });
 
 class postController {
   //* 과제 3. 게시물 생성
-  static async createPost(req: Request, res: Response, next: NextFunction) {
+  static async createPost(
+    req: UserIdReqType,
+    res: Response,
+    next: NextFunction,
+  ) {
+    const userId = req.currentUserId;
     try {
       const { title, description } = await postReqValidator.validateAsync(
         req.body,
       );
-      const post = await postService.savePost({ title, description });
+      const post = await postService.savePost({ userId, title, description });
       return res.status(201).send(post);
     } catch (error) {
       next(error);
@@ -43,24 +53,45 @@ class postController {
     }
   }
   //* 과제 6. 특정 게시글 수정
-  static async editPost(req: Request, res: Response, next: NextFunction) {
+  static async editPost(req: UserIdReqType, res: Response, next: NextFunction) {
     try {
+      const userId = req.currentUserId;
       const { id } = req.params as PostParamsType;
       const { title, description } = await postReqValidator.validateAsync(
         req.body,
       );
-      const postInfo: PostInfoType = { id, title, description };
+
+      const postInfo: PostInfoType = {
+        id,
+        userId,
+        title,
+        description,
+      };
       const post = await postService.updatePost(postInfo);
+
+      if (typeof post === 'string') {
+        throw ApiError.setBadRequest(post);
+      }
       return res.status(200).send(post);
     } catch (error) {
       next(error);
     }
   }
   //* 과제 7. 특정 게시글 삭제
-  static async deletePost(req: Request, res: Response, next: NextFunction) {
+  static async deletePost(
+    req: UserIdReqType,
+    res: Response,
+    next: NextFunction,
+  ) {
     try {
+      const userId = req.currentUserId;
       const { id } = req.params as unknown as PostParamsType;
-      const post = await postService.removePost({ id });
+      const post = await postService.removePost({ id, userId });
+
+      if (typeof post === 'string') {
+        throw ApiError.setBadRequest(post);
+      }
+
       return res.status(200).send(post);
     } catch (error) {
       next(error);
